requirements.txt
jsf==0.10.0
jsonpointer==2.4
jsonschema==4.21.1
python-benedict==0.33.1
